Add `semi_focused` tags
I use multiple outputs, and as such often need to remember which
output is currently focused.

Fortunately, river tells us which output is currently in focus.
Therefore, I added the `is_semi_focused` flag to a `Tag`, to signify
that it is a focused Tag on its output, but that it is _not_ actually
focused, because the whole output is not focused.

For users that do not need or want this extra differentiation, setting
`tag_semi_focused_{fg,bg}` to the `tag_focused_{fg,bg}` values, will
result in the previous behaviour.

I am not familiar enough with niri or hyprland to know if they can
support something like this too. As such, both of their information
providers set `is_semi_focused` unconditionally to `false`.