fix(crate::wm_info_provider::river): Re-calculate the tags on output focus change
This was previously removed, in favour of only re-drawing the bar, but
is actually needed, as otherwise the different colours for the tags
(`tag_{,focused,...}_{fg,bg}`) don't take effect at all.

I think, that a better option, would be to simply allow redrawing the
tags, but avoiding re-calculating them.